Move structures to own directory
