+++ base_util.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <stdbool.h>
-#include "structures.h"
+#include "structures/structures.h"
 
 /* Macro because it makes printf errors easier to detect at compile time */
 /* Minor hack because c99 doesn't allow zero-length varargs for macros */
@@ -28,26 +28,10 @@
     }\
 } while (0)
 
-/* C compile-time constant madness*/
-#define STRLEN(s) ( (sizeof(s)/sizeof(s[0])) - sizeof(s[0]) )
-#define STR_NEW(str)\
-{\
-        .length = STRLEN(str),\
-        .body=(str)\
-}
-/*
-struct String string_new(char *str);
-*/
-bool string_is_sane(struct String *s);
-struct String string_copy_stack(struct String *from);
-struct String *copy_string_heap(struct String *from);
-void string_copy_to(struct String *to, struct String *from);
-void string_concat_to(struct String *to, struct String *from);
 int string_contains(char ch, struct String* string);
 struct Value copy_value_stack(struct Value * b);
 struct Value *copy_value_heap(struct Value * b);
 void copy_value_to(struct Value * a, struct Value * b);
-int string_matches(struct String *base, struct String *compare);
 void printTree(struct Tree *tree);
 struct Tree * duplicate_tree(struct Tree * a);
 void copy_array(struct Value * a, struct Value_array * b);
