+++ build_array.h
@@ -1,7 +1,7 @@
 #ifndef _BUILD_ARRAY_H
 #define _BUILD_ARRAY_H
 
-#include "structures.h"
+#include "structures/structures.h"
 
 void build_array(struct Value_array * array, struct Tokens * tokens);
