+++ execution_functions/for_each.h
@@ -1,7 +1,7 @@
 #ifndef _FOR_EACH_H
 #define _FOR_EACH_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value for_each(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
