+++ execution_functions/for_loop.h
@@ -1,7 +1,7 @@
 #ifndef _FOR_LOOP_H
 #define _FOR_LOOP_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 void for_loop(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
