+++ execution_functions/if_block.h
@@ -1,7 +1,7 @@
 #ifndef _IF_BLOCK_H
 #define _IF_BLOCK_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value if_block(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
