+++ execution_functions/let_binding.h
@@ -1,7 +1,7 @@
 #ifndef _LET_BINDING_H
 #define _LET_BINDING_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 void store_let_binding(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
