+++ execution_functions/map_array.c
@@ -1,4 +1,4 @@
-#include "../structures.h"
+#include "../structures/structures.h"
 #include "../execute.h"
 #include "../base_util.h"
 #include "populate_each_function.h"
