+++ execution_functions/map_array.h
@@ -1,7 +1,7 @@
 #ifndef _MAP_ARRAY_H
 #define _MAP_ARRAY_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value map_array(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
