+++ execution_functions/populate_each_function.h
@@ -1,7 +1,7 @@
 #ifndef _POPULATE_EACH_FUNCTION_H
 #define _POPULATE_EACH_FUNCTION_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 void populate_each_function(struct Value *, struct Value *, struct Tree *, struct Value *, struct Value *);
