+++ execution_functions/read_file.h
@@ -1,7 +1,7 @@
 #ifndef _READ_FILE_H
 #define _READ_FILE_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value read_file(struct String);
