+++ execution_functions/reduce.h
@@ -1,7 +1,7 @@
 #ifndef _REDUCE_H
 #define _REDUCE_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value reduce_array(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
