+++ execution_functions/switch.h
@@ -1,7 +1,7 @@
 #ifndef _SWITCH_H
 #define _SWITCH_H
 
-#include "../structures.h"
+#include "../structures/structures.h"
 
 struct Value switch_case(struct Tree * ast, struct Tree_map * defined, struct Map * let_map, int max_depth);
