+++ parser.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
-#include "structures.h"
+#include "structures/structures.h"
 #include "constants.h"
 #include "core_functions.h"
 #include "build_array.h"
