+++ parser.h
@@ -1,7 +1,7 @@
 #ifndef _PARSER_H
 #define _PARSER_H
 
-#include "structures.h"
+#include "structures/structures.h"
 
 void parse(struct Tree * root, struct Tokens *tokens);
