+++ process_defs.h
@@ -1,7 +1,7 @@
 #ifndef _PROCESS_DEFS_H
 #define _PROCESS_DEFS_H
 
-#include "structures.h"
+#include "structures/structures.h"
 
 struct Map * make_args_map(struct Tree * ast, struct Tree_map * defined, struct Map *, int idx, int max_depth);
 struct Tree * get_defined_body(struct Tree * function);
