+++ tokenizer.h
@@ -1,7 +1,7 @@
 #ifndef _TOKENIZER_H
 #define _TOKENIZER_H
 
-#include "structures.h"
+#include "structures/structures.h"
 
 struct Tokens * tokenize(struct String file, struct Tokens *content);
